Crate tt

tt crate defines a TokenTree data structure: this is the interface (both input and output) of macros. It closely mirrors proc_macro crate’s TokenTree.

Modules

buffer

Stateful iteration over token trees.

Structs

DelimSpan

Delimiter

Ident

Identifier or keyword. Unlike rustc, we keep “r#” prefix when it represents a raw identifier.

Literal

Punct

SmolStr

A SmolStr is a string type that has the following properties:

Subtree

SubtreeBuilder

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

Enums

DelimiterKind

Leaf

Spacing

TokenTree

Functions

pretty