tt

Module storage

Spans are memory heavy, and we have a lot of token trees. Storing them straight will waste a lot of memory. So instead we implement a clever compression mechanism:

A TopSubtree has a list of CompressedSpanPart, which are the parts of a span that tend to be shared between tokens - namely, without the range.

The main list of token trees is stored in a variable-length encoding as bytes. The encoding is documented in the decode() function (which decodes one TokenTree).

Structs

BufferReader 🔒

BufferWriter 🔒

CompressedSpanPart 🔒

InlineSpanParts 🔒

SubtreeBuilderRestorePoint

TokenTreesSlice 🔒

TopSubtree

TopSubtreeBuilder

UninitBuffer 🔒

Constants

BIGGEST_POSSIBLE_TT_ENCODING 🔒

This is subtree in format 4: two spans, each at most 3*u32, a u8 control byte, a u32 length and a u32 bytes length.

Traits

Encodable 🔒

Functions

change_root_delimiter 🔒

change_root_spans 🔒 ⚠

compute_span_frequencies_and_symbols 🔒

decode 🔒 ⚠

We need MaybeUninit to preserve provenance.

decode_extended_span 🔒 ⚠

decode_span 🔒 ⚠

decode_symbol 🔒 ⚠

encode 🔒 ⚠

encode_all 🔒

encode_extended_span 🔒 ⚠

encode_span 🔒 ⚠

encode_span_no_map 🔒 ⚠

encode_symbol 🔒 ⚠

encode_top_subtree 🔒 ⚠

We always encode the top subtree with the heaviest encoding because we sometimes want to change it.

n_bits_mask 🔒

n_bits_mask_u64 🔒