tt

Module storage

Spans are memory heavy, and we have a lot of token trees. Storing them straight will waste a lot of memory. So instead we implement a clever compression mechanism:

A TopSubtree has a list of CompressedSpanPart, which are the parts of a span that tend to be shared between tokens - namely, without the range. The main list of token trees is kept in one of three versions, where we use the smallest version we can for this tree:

In the most common version a span is just a u32. The bits are divided as follows: there are 4 bits that index into the CompressedSpanPart list. 20 bits store the range start, and 8 bits store the range length. In experiments, this accounts for 75%-85% of the spans.

In the second version a span is 64 bits. 32 bits for the range start, 16 bits for the range length, and 16 bits for the span parts index. This is used in less than 2% of all TopSubtrees, but they account for 15%-25% of the spans: those are mostly token tree munchers, that generate a lot of SyntaxContexts (because they recurse a lot), which is why they can’t fit in the first version, and tend to generate a lot of code.

The third version is practically unused; 65,535 bytes for a token and 65,535 unique span parts is more than enough for everybody. However, someone may still create a macro that requires more, therefore we have this version as a backup: it uses 96 bits, 32 for each of the range start, length and span parts index.

Macros

dispatch 🔒

dispatch_builder 🔒

Structs

CompressedSpanPart 🔒

SpanStorage32 🔒

SpanStorage64 🔒

SpanStorage96 🔒

SubtreeBuilderRestorePoint

TopSubtree

TopSubtreeBuilder

Enums

TokenTree 🔒

TopSubtreeBuilderRepr 🔒

TopSubtreeRepr 🔒

Constants

RESERVED_SPAN_PARTS_LEN 🔒

In any tree, the first two subtree parts are reserved for the top subtree.

Traits

SpanStorage 🔒

Functions

n_bits_mask 🔒

Type Aliases

FxIndexSet 🔒